# Speeding up Scala 3 Writing macros