--- a/build.sc
+++ b/build.sc
@@ -18,7 +18,7 @@ import com.github.lolgab.mill.mima._
 val scala211  = "2.11.12"
 val scala212  = "2.12.17"
 val scala213  = "2.13.10"
-val scala3   = "3.1.3"
+val scala3   = "3.2.1"
 val scalaJS  = "1.10.1"
 val scalaNative = "0.4.5"
 val acyclic = "0.3.6"