--- a/implicits/src-3/upickle/implicits/CaseClassWriter.scala
+++ b/implicits/src-3/upickle/implicits/CaseClassWriter.scala
@@ -72,8 +72,7 @@ trait CaseClassWriterPiece extends MacrosCommon:
             thisOuter,
             this,
             v,
-            ctx,
-            macros.summonList[Tuple.Map[m.MirroredElemTypes, Writer]]
+            ctx
           )
       }