--- a/implicits/src-3/upickle/implicits/CaseClassWriter.scala
+++ b/implicits/src-3/upickle/implicits/CaseClassWriter.scala
@@ -8,36 +8,6 @@ import upickle.core.{ Visitor, ObjVisitor, Annotator }
 trait CaseClassWriterPiece extends MacrosCommon:
   this: upickle.core.Types with Writers with Annotator =>
   transparent inline def thisOuter = this
-  class CaseClassWriter[V](
-    elemsInfo: V => List[(String, Writer[_], Any)],
-    defaultParams: Map[String, AnyRef]) extends CaseW[V]:
-    def length(v: V): Int =
-      var n = 0
-      for
-        (name, _, value) <- elemsInfo(v)
-        defaultValue = defaultParams.get(name)
-        if serializeDefaults || defaultValue.isEmpty || defaultValue.get != value
-      do n += 1
-      n
-    end length
-
-    def writeToObject[R](ctx: ObjVisitor[_, R], v: V): Unit =
-      for
-        (name, writer, value) <- elemsInfo(v)
-        defaultValue = defaultParams.get(name)
-        if serializeDefaults || defaultValue.isEmpty || defaultValue.get != value
-      do
-        val keyVisitor = ctx.visitKey(-1)
-        ctx.visitKeyValue(
-          keyVisitor.visitString(
-            objectAttributeKeyWriteMap(name),
-            -1
-          )
-        )
-        ctx.narrow.visitValue(
-          writer.narrow.write(ctx.subVisitor, value), -1)
-    end writeToObject
-  end CaseClassWriter
 
   class EnumWriter[T] extends Writer[T]:
     override def write0[V](out: Visitor[_, V], v: T): V = out.visitString(v.toString, -1)
@@ -47,25 +17,26 @@ trait CaseClassWriterPiece extends MacrosCommon:
     case m: Mirror.ProductOf[T] =>
       def writer = new CaseW[T] {
         def length(v: T) =
-          def elemsInfo(v: T): List[(String, Writer[_], Any)] =
-            val labels: List[String] = macros.fieldLabels[T].map(_._1)
-            val writers: List[Writer[_]] =
-              macros.summonList[Tuple.Map[m.MirroredElemTypes, Writer]]
-                .asInstanceOf[List[Writer[_]]]
-            val values: List[Any] = v.asInstanceOf[Product].productIterator.toList
-            for ((l, w), v) <- labels.zip(writers).zip(values)
-              yield (l, w, v)
-
-          val defaultParams =
-            macros.getDefaultParams[T]
-
-          var n = 0
-          for
-            (name, _, value) <- elemsInfo(v)
-            defaultValue = defaultParams.get(name)
-            if serializeDefaults || defaultValue.isEmpty || defaultValue.get != value
-          do n += 1
-          n
+          macros.writeLength(v)
+//          def elemsInfo(v: T): List[(String, Writer[_], Any)] =
+//            val labels: List[String] = macros.fieldLabels[T].map(_._1)
+//            val writers: List[Writer[_]] =
+//              macros.summonList[Tuple.Map[m.MirroredElemTypes, Writer]]
+//                .asInstanceOf[List[Writer[_]]]
+//            val values: List[Any] = v.asInstanceOf[Product].productIterator.toList
+//            for ((l, w), v) <- labels.zip(writers).zip(values)
+//              yield (l, w, v)
+//
+//          val defaultParams =
+//            macros.getDefaultParams[T]
+//
+//          var n = 0
+//          for
+//            (name, _, value) <- elemsInfo(v)
+//            defaultValue = defaultParams.get(name)
+//            if serializeDefaults || defaultValue.isEmpty || defaultValue.get != value
+//          do n += 1
+//          n
 
         def writeToObject[R](ctx: _root_.upickle.core.ObjVisitor[_, R], v: T): Unit =
           macros.writeSnippets[R, T, Tuple.Map[m.MirroredElemTypes, Writer]](