… implicits/src-3/upickle/implicits/Writers.scala
… implicits/src-3/upickle/implicits/macros.scala
… upickle/test/src-2/upickle/AdvancedTestsScala2Only.scala
… upickle/test/src-3/upickle/Derivation.scala
… upickle/test/src/upickle/AdvancedTests.scala