Fixes https://github.com/com-lihaoyi/upickle/issues/389 and https://github.com/com-lihaoyi/upickle/issues/388

This PR moves a bunch of logic from runtime to compile time, and optimizes whatever runtime logic remains. It also consolidates the logic with the Scala 2 logic as much as possible, e.g. re-using `writeSnippet` to manage writes and  `CaseObjectContext`/`HugeCaseObjectContext` to manage validation and error reporting during a read

Running some ad-hoc benchmarks on my laptop, `./mill bench.jvm.run`, it gives a significant speedup on reads and writes, and brings it close to the Scala 2.13 numbers (higher is better):

| Benchmark (5000ms)                  | Scala 3 Before | Scala 3 After | Scala 2 |      
|-------------------------------------|----------------|---------------|---------|      
| upickleDefault Read                 |      637       |     1065      |   1403  |      
| upickleDefault Write                |      839       |     1452      |   1549  |      
| upickleDefaultByteArray Read        |      582       |     1172      |   1126  |      
| upickleDefaultByteArray Write       |      847       |     1218      |   1277  |      
| upickleDefaultBinary Read           |      925       |     3853      |   3844  |      
| upickleDefaultBinary Write          |      1252      |     3117      |   3666  |      
| upickleDefaultCached Read           |      620       |     1300      |   1412  |      
| upickleDefaultCached Write          |      829       |     1555      |   1588  |      
| upickleDefaultByteArrayCached Read  |      575       |     1182      |   1095  |      
| upickleDefaultByteArrayCached Write |      838       |     1223      |   1297  |      
| upickleDefaultBinaryCached Read     |      928       |     3825      |   3885  |      
| upickleDefaultBinaryCached Write    |      1266      |     2907      |   3674  |      

Note that the generated code, especially for reads, is still not as optimized as the Scala 2 versions:

1. I couldn't figure out how to generate an anonymous class with typed fields in Scala 3, so I'm putting things in an `Array[Any]`
2. I couldn't figure out how to generate `match` statements, so I generated a `Map[K, V]` and look it up at runtime.

Fixing this issues and moving the reading logic into the macro should also be possible, but can happen in a separate PR

All existing tests pass. Added a regression test for the recursive Scala 3 scenario that hangs on master. Also moved a bunch of `AdvancedTests` into the shared folder now that they work