@lihaoyi 2023-02-02
I'll leave this over the weekend, and if there aren't any more comments I'll merge it then. All the existing/new/newly-enabled tests pass, and the benchmarks demonstrate the non-functional improvements, so I can't have broken anything too badly

The issue noticed by @bishabosha is a limitation in existing code, which is why it was not picked up by any tests (since it never worked). Fixing that can come in a follow up