@lefou 2023-02-06
There seem to be issues with this change. E.g. in Ammonite, the integration tests don't succeed. 

* https://github.com/com-lihaoyi/Ammonite/pull/1320